Deprecate API in Web.Twitter.Conduit.Status
